requirements.txt
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
requests>=2.31.0
pyperclip>=1.8.2
pywin32>=306; sys_platform == 'win32'

